feat(exporter/clickhouse): Use new JSON type for attributes
New JSON type is now marked as stable in the latest releases of
Clickhouse and is recommended for use. This commit replaces all
columns with type Map(LowCardinality(String) String) with JSON and
removes the indexed on the keys and values, since they aren't necessary
anymore. JSON is supposed to give efficiency benefits while storing and
querying columns with arbitrary values stored as key-value pairs.

Further optimisation that can be done is providing a dedicated
Serializer for the clickhouse-go client for the said columns while
inserting.